Fixed issue #85.

+++ biojava3-alignment/src/main/java/org/biojava3/alignment/FractionalSimilarityScorer.java
@@ -87,7 +87,7 @@ public class FractionalSimilarityScorer<S extends Sequence<C>, C extends Compoun
     // methods for Scorer
 
     @Override
-    public int getMaxScore() {
+    public double getMaxScore() {
         if (aligner != null) {
             align();
         }
@@ -95,12 +95,12 @@ public class FractionalSimilarityScorer<S extends Sequence<C>, C extends Compoun
     }
 
     @Override
-    public int getMinScore() {
+    public double getMinScore() {
         return 0;
     }
 
     @Override
-    public int getScore() {
+    public double getScore() {
         if (aligner != null) {
             align();
         }
