+++ biojava3-alignment/src/main/java/org/biojava3/alignment/template/AbstractMatrixAligner.java
@@ -240,7 +240,7 @@ public abstract class AbstractMatrixAligner<S extends Sequence<C>, C extends Com
     // methods for Scorer
 
     @Override
-    public int getMaxScore() {
+    public double getMaxScore() {
         if (profile == null) {
             align();
         }
@@ -248,7 +248,7 @@ public abstract class AbstractMatrixAligner<S extends Sequence<C>, C extends Com
     }
 
     @Override
-    public int getMinScore() {
+    public double getMinScore() {
         if (profile == null) {
             align();
         }
@@ -256,7 +256,7 @@ public abstract class AbstractMatrixAligner<S extends Sequence<C>, C extends Com
     }
 
     @Override
-    public int getScore() {
+    public double getScore() {
         if (profile == null) {
             align();
         }
